djangorequirements.txt
psycopg2-binary==2.9.1
Django~=3.2.6
django-environ==0.4.5
djangorestframework~=3.12.4
django-rest-swagger==2.2.0
sphinx==4.0.2
drf-yasg==1.20.0
gunicorn==20.1.0
drf-jwt==1.19.1
drf-api-logger==1.0.8
djangorestframework-simplejwt==4.8.0
django-rest-passwordreset==1.2.0
rest_condition==1.0.3
django-dry-rest-permissions==1.2.0
django-simple-history~=3.0.0
drf-api-tracking==1.8.0
coreapi~=2.3.3
python-decouple==3.4
PyJWT==2.1.0
pandas==1.3.2
pyarrow
django-utils-six
sqlalchemy
openpyxl
django-cors-headers